# Reduce fixed overhead of some Utf8Parser.TryParse methods