--- a/src/libraries/System.Private.CoreLib/src/System/Buffers/Text/Utf8Parser/ParserHelpers.cs
+++ b/src/libraries/System.Private.CoreLib/src/System/Buffers/Text/Utf8Parser/ParserHelpers.cs
@@ -5,6 +5,7 @@
 using System.Diagnostics;
 using System.Diagnostics.CodeAnalysis;
 using System.Runtime.CompilerServices;
+using Internal.Runtime.CompilerServices;
 
 namespace System.Buffers.Text
 {
@@ -80,15 +81,15 @@ namespace System.Buffers.Text
         [StackTraceHidden]
         public static bool TryParseThrowFormatException<T>(ReadOnlySpan<byte> source, out T value, out int bytesConsumed) where T : struct
         {
-            // While loop below tells C# compiler to ignore 'out not assigned' errors
-            // and tells JIT that this method never returns. The parameters to this method
-            // are ordered the same as our callers' parameters so as to allow the JIT to
-            // avoid unnecessary register swapping.
+            // The parameters to this method are ordered the same as our callers' parameters
+            // allowing the JIT to avoid unnecessary register swapping or spilling.
 
-            while (true)
-            {
-                ThrowHelper.ThrowFormatException_BadFormatSpecifier();
-            }
+            Unsafe.SkipInit(out value); // bypass language initialization rules since we're about to throw
+            Unsafe.SkipInit(out bytesConsumed);
+            ThrowHelper.ThrowFormatException_BadFormatSpecifier();
+
+            Debug.Fail("Control should never reach this point.");
+            return false;
         }
     }
 }