--- a/src/libraries/System.Private.CoreLib/src/System/Buffers/Text/Utf8Parser/Utf8Parser.Integer.Signed.cs
+++ b/src/libraries/System.Private.CoreLib/src/System/Buffers/Text/Utf8Parser/Utf8Parser.Integer.Signed.cs
@@ -35,26 +35,28 @@ namespace System.Buffers.Text
         [CLSCompliant(false)]
         public static bool TryParse(ReadOnlySpan<byte> source, out sbyte value, out int bytesConsumed, char standardFormat = default)
         {
-            switch (standardFormat)
+        FastPath:
+            if (standardFormat == default)
             {
-                case default(char):
+                return TryParseSByteD(source, out value, out bytesConsumed);
+            }
+
+            switch (standardFormat | 0x20) // convert to lowercase
+            {
                 case 'g':
-                case 'G':
                 case 'd':
-                case 'D':
-                    return TryParseSByteD(source, out value, out bytesConsumed);
+                    standardFormat = default;
+                    goto FastPath;
 
                 case 'n':
-                case 'N':
                     return TryParseSByteN(source, out value, out bytesConsumed);
 
                 case 'x':
-                case 'X':
-                    value = default;
+                    Unsafe.SkipInit(out value); // will be populated by TryParseByteX
                     return TryParseByteX(source, out Unsafe.As<sbyte, byte>(ref value), out bytesConsumed);
 
                 default:
-                    return ParserHelpers.TryParseThrowFormatException(out value, out bytesConsumed);
+                    return ParserHelpers.TryParseThrowFormatException(source, out value, out bytesConsumed);
             }
         }
 
@@ -81,26 +83,28 @@ namespace System.Buffers.Text
         /// </exceptions>
         public static bool TryParse(ReadOnlySpan<byte> source, out short value, out int bytesConsumed, char standardFormat = default)
         {
-            switch (standardFormat)
+        FastPath:
+            if (standardFormat == default)
             {
-                case default(char):
+                return TryParseInt16D(source, out value, out bytesConsumed);
+            }
+
+            switch (standardFormat | 0x20) // convert to lowercase
+            {
                 case 'g':
-                case 'G':
                 case 'd':
-                case 'D':
-                    return TryParseInt16D(source, out value, out bytesConsumed);
+                    standardFormat = default;
+                    goto FastPath;
 
                 case 'n':
-                case 'N':
                     return TryParseInt16N(source, out value, out bytesConsumed);
 
                 case 'x':
-                case 'X':
-                    value = default;
+                    Unsafe.SkipInit(out value); // will be populated by TryParseUInt16X
                     return TryParseUInt16X(source, out Unsafe.As<short, ushort>(ref value), out bytesConsumed);
 
                 default:
-                    return ParserHelpers.TryParseThrowFormatException(out value, out bytesConsumed);
+                    return ParserHelpers.TryParseThrowFormatException(source, out value, out bytesConsumed);
             }
         }
 
@@ -127,26 +131,28 @@ namespace System.Buffers.Text
         /// </exceptions>
         public static bool TryParse(ReadOnlySpan<byte> source, out int value, out int bytesConsumed, char standardFormat = default)
         {
-            switch (standardFormat)
+        FastPath:
+            if (standardFormat == default)
             {
-                case default(char):
+                return TryParseInt32D(source, out value, out bytesConsumed);
+            }
+
+            switch (standardFormat | 0x20) // convert to lowercase
+            {
                 case 'g':
-                case 'G':
                 case 'd':
-                case 'D':
-                    return TryParseInt32D(source, out value, out bytesConsumed);
+                    standardFormat = default;
+                    goto FastPath;
 
                 case 'n':
-                case 'N':
                     return TryParseInt32N(source, out value, out bytesConsumed);
 
                 case 'x':
-                case 'X':
-                    value = default;
+                    Unsafe.SkipInit(out value); // will be populated by TryParseUInt32X
                     return TryParseUInt32X(source, out Unsafe.As<int, uint>(ref value), out bytesConsumed);
 
                 default:
-                    return ParserHelpers.TryParseThrowFormatException(out value, out bytesConsumed);
+                    return ParserHelpers.TryParseThrowFormatException(source, out value, out bytesConsumed);
             }
         }
 
@@ -173,26 +179,28 @@ namespace System.Buffers.Text
         /// </exceptions>
         public static bool TryParse(ReadOnlySpan<byte> source, out long value, out int bytesConsumed, char standardFormat = default)
         {
-            switch (standardFormat)
+        FastPath:
+            if (standardFormat == default)
             {
-                case default(char):
+                return TryParseInt64D(source, out value, out bytesConsumed);
+            }
+
+            switch (standardFormat | 0x20) // convert to lowercase
+            {
                 case 'g':
-                case 'G':
                 case 'd':
-                case 'D':
-                    return TryParseInt64D(source, out value, out bytesConsumed);
+                    standardFormat = default;
+                    goto FastPath;
 
                 case 'n':
-                case 'N':
                     return TryParseInt64N(source, out value, out bytesConsumed);
 
                 case 'x':
-                case 'X':
-                    value = default;
+                    Unsafe.SkipInit(out value); // will be populated by TryParseUInt64X
                     return TryParseUInt64X(source, out Unsafe.As<long, ulong>(ref value), out bytesConsumed);
 
                 default:
-                    return ParserHelpers.TryParseThrowFormatException(out value, out bytesConsumed);
+                    return ParserHelpers.TryParseThrowFormatException(source, out value, out bytesConsumed);
             }
         }
     }