--- a/src/libraries/System.Private.CoreLib/src/System/Buffers/Text/Utf8Parser/Utf8Parser.Integer.Unsigned.cs
+++ b/src/libraries/System.Private.CoreLib/src/System/Buffers/Text/Utf8Parser/Utf8Parser.Integer.Unsigned.cs
@@ -35,6 +35,9 @@ namespace System.Buffers.Text
                 return TryParseByteD(source, out value, out bytesConsumed);
             }
 
+            // There's small but measurable overhead when entering the switch block below.
+            // We optimize for the default case by hoisting it above the switch block.
+
             switch (standardFormat | 0x20) // convert to lowercase
             {
                 case 'g':
@@ -82,6 +85,9 @@ namespace System.Buffers.Text
             {
                 return TryParseUInt16D(source, out value, out bytesConsumed);
             }
+
+            // There's small but measurable overhead when entering the switch block below.
+            // We optimize for the default case by hoisting it above the switch block.
 
             switch (standardFormat | 0x20) // convert to lowercase
             {
@@ -131,6 +137,9 @@ namespace System.Buffers.Text
                 return TryParseUInt32D(source, out value, out bytesConsumed);
             }
 
+            // There's small but measurable overhead when entering the switch block below.
+            // We optimize for the default case by hoisting it above the switch block.
+
             switch (standardFormat | 0x20) // convert to lowercase
             {
                 case 'g':
@@ -178,6 +187,9 @@ namespace System.Buffers.Text
             {
                 return TryParseUInt64D(source, out value, out bytesConsumed);
             }
+
+            // There's small but measurable overhead when entering the switch block below.
+            // We optimize for the default case by hoisting it above the switch block.
 
             switch (standardFormat | 0x20) // convert to lowercase
             {