@stephentoub 2020-03-12
How does this relate to https://github.com/dotnet/runtime/pull/32843?

Can similar tweaks be made to the corresponding char-based APIs (e.g. int.Parse)?