@GrabYourPitchforks 2020-03-12
@stephentoub That PR was primarily concerned with improving the workhorse routine. From that investigation came the idea of improving the entry point. This PR focuses just on the fixed overhead of the entry points while leaving the workhorse routines largely unchanged.

I haven't tried applying these optimizations to the char overloads yet because their entry points generally look different than these entry points. But I imagine once we get back to the PR that optimizes the workhorse routines the optimizations there will be generally applicable to both char and byte.